Move build to vite; test to vitest; upgrade deps
