+++ package.json
@@ -19,6 +19,7 @@
     },
     "./client": {
       "import": "./dist/client/index.js",
+      "require": "./dist/client/index.cjs",
       "types": "./dist/client/index.d.ts"
     },
     "./virtual": {
@@ -37,23 +38,22 @@
   "module": "dist/node/index.js",
   "types": "dist/node/index.d.ts",
   "scripts": {
-    "build": "node --experimental-loader esbuild-node-loader scripts/build.ts",
-    "test": "node --experimental-loader esbuild-node-loader --experimental-specifier-resolution=node ./node_modules/.bin/uvu tests",
+    "build": "node --experimental-loader esbuild-node-loader scripts/build.ts && tsc -p tsconfig.build.json",
+    "test": "vitest --globals",
     "prepublish": "npm run build"
   },
   "dependencies": {
-    "@remix-run/dev": "^1.0.6"
+    "@remix-run/dev": "^1.3.3"
   },
   "devDependencies": {
     "@types/react": "16.7.0",
-    "esbuild-node-loader": "^0.6.3",
+    "esbuild-node-loader": "^0.8.0",
     "react": "16.6.0",
     "react-router": "6.0.0",
     "react-router-dom": "6.0.0",
-    "tsup": "^5.10.0",
     "typescript": "^4.5.2",
-    "uvu": "^0.5.2",
-    "vite": "^2.6.14"
+    "vite": "^2.8.6",
+    "vitest": "^0.7.10"
   },
   "peerDependencies": {
     "react": "^16.6.0",
