+++ scripts/build.ts
@@ -1,28 +1,34 @@
 import path from "node:path";
-import url from "node:url";
-import fs from "node:fs/promises";
 
-import tsup from "tsup";
+import { build, LibraryFormats } from "vite";
 
-const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
+const fileName = (format: string) => `index.${format === "es" ? "js" : "cjs"}`;
+const formats: LibraryFormats[] = ["es", "cjs"];
 
-const dist = path.join(__dirname, "..", "dist");
-
-await fs
-  .access(dist)
-  .then(() => fs.rm(dist, { recursive: true }))
-  .catch(() => {});
-
-await tsup.build({
-  entryPoints: ["lib/node/index.ts"],
-  format: ["esm", "cjs"],
-  dts: true,
-  outDir: "dist/node",
+await build({
+  build: {
+    outDir: "dist/node",
+    lib: {
+      entry: path.resolve(process.cwd(), "lib/node/index.ts"),
+      formats,
+      fileName,
+    },
+    rollupOptions: {
+      external: [/@remix-run\/dev\/.+/g, "vite"],
+    },
+  },
 });
 
-await tsup.build({
-  entryPoints: ["lib/client/index.ts"],
-  format: ["esm"],
-  dts: true,
-  outDir: "dist/client",
+await build({
+  build: {
+    outDir: "dist/client",
+    lib: {
+      entry: path.resolve(process.cwd(), "lib/client/index.ts"),
+      formats,
+      fileName,
+    },
+    rollupOptions: {
+      external: ["react", "react-router-dom"],
+    },
+  },
 });
