+++ tsconfig.build.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": false,
+    "outDir": "dist",
+    "emitDeclarationOnly": true
+  },
+  "include": ["lib"]
+}
